As expression separated
